chore: open-source release prep — docs restructure, narrative, trust files
Restructure the repo for an open-source release without changing runtime
behavior.

Narrative
- Reframe to "Vibe first. Then FastAgent.": take a local agent folder out of
  the terminal and serve it in an app, on GitHub, in Telegram, or behind a
  custom channel.
- README: features, design philosophy, Why FastAgent, badges, project links.

Docs
- Add docs/README.md index, overview, cli, configuration, api-reference,
  troubleshooting, principles, and dedicated channel guides (github, telegram,
  channel-development).
- Move maintainer notes under docs/design/ (core.md, README.md).
- Remove private strategy docs (positioning, comparisons) and the old
  fastagent.md / core-design.md.

Trust / community files
- Add SECURITY.md, CHANGELOG.md, llms.txt, and start.md (AI-guided setup).

Code housekeeping (no behavior change)
- Move engines/pi/{init,add-channel,vendor-skill,scaffold-templates}.ts into
  engines/pi/scaffold/; rename dev.ts -> workspace.ts.
- Drop core/examples/ and its tsconfig/biome/package references (better
  examples will be added later).
- Update imports/exports/tests for the moves.
