+++ CHANGELOG.md
@@ -1,0 +1,32 @@
+# Changelog
+
+All notable changes to `@kid7st/fastagent` are documented here.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+While the project is pre-1.0, minor versions may include breaking changes.
+
+## [Unreleased]
+
+### Added
+- `SECURITY.md`, `CHANGELOG.md`, and `llms.txt` for open-source readiness.
+- README status badges (CI, npm version, license, Node version).
+- `start.md`: an AI-guided setup prompt for coding agents.
+- Restructured public docs: `docs/README.md` index, `overview`, `cli`, `configuration`,
+  `api-reference`, `troubleshooting`, `principles`, dedicated channel guides
+  (`github`, `telegram`, `channel-development`), and maintainer notes under `docs/design/`.
+
+### Changed
+- Narrative reframed to "Vibe first. Then FastAgent." — take a local agent folder out of
+  the terminal and serve it in an app, on GitHub, in Telegram, or behind a custom channel.
+
+### Removed
+- `core/examples/` (better examples will be added later).
+
+## [0.7.1]
+
+Last release before the open-source documentation pass. Earlier history is in the
+[commit log](https://github.com/kid7st/fastagent/commits/main).
+
+[Unreleased]: https://github.com/kid7st/fastagent/compare/v0.7.1...HEAD
+[0.7.1]: https://github.com/kid7st/fastagent/releases/tag/v0.7.1
