--- core/examples/agent/skills/refund-policy/SKILL.md
@@ -1,10 +1,0 @@
----
-name: refund-policy
-description: Use when a support message asks about refunds. Contains the canonical refund policy.
----
-
-# Refund policy (canonical)
-
-- Full refund within 14 days of purchase, no questions asked.
-- 50% refund within 30 days.
-- After 30 days: no refund, but offer a 20% discount coupon (code SORRY20).
