+++ core/tsconfig.json
@@ -13,5 +13,5 @@
     "verbatimModuleSyntax": true,
     "allowImportingTsExtensions": true
   },
-  "include": ["src", "test", "examples"]
+  "include": ["src", "test"]
 }
